extract MenuItemCard into separate component
