+++ src/dataTypes.tsx
@@ -1,0 +1,7 @@
+export interface MenuItem {
+  imageUrl: string
+  imageDescription: string
+  name: string
+  price: string
+  description: string
+}
